Move FINUFFTBackend to package extension
This is mainly to avoid the heavy dependency of FINUFFT.jl on CUDA.jl,
which is not necessarily used by all backends.
